Extract k6/experimental/browser.Page definitions
